+++ db/migrate/20210214141307_add_default_payment_info_to_organisations.rb
@@ -1,0 +1,6 @@
+class AddDefaultPaymentInfoToOrganisations < ActiveRecord::Migration[6.1]
+  def change
+    add_column :organisations, :default_payment_info_type, :string
+    add_column :organisations, :invoice_ref_template, :string, default: '%<prefix>s%<home_id>03d%<tenant_id>06d%<invoice_id>07d'
+  end
+end